Develop - 2ºnd "Release" polishing and "better code" overall (#2)
* deleted redundant booky.py file

* refactor: removed workaround for default arg
- implemented it directly into the main funcion

* refactor: path as default arg, added comments
- Path is now the default arg only when nothing else ran

* fixed errors when running optional arguments
- without anything else

* refactor: shared read call to json file
- so that the script does'nt have to read as much from disk.

* refactor: better readability and comments
- Fixed --remove option removing the whole file, too!

+++ src/booky/booky.py
@@ -7,7 +7,6 @@
 """
 import argparse
 import json
-# import toml
 from pathlib import Path
 import sys
 import os
@@ -20,6 +19,7 @@ def argParser() -> argparse.ArgumentParser:
     # Manage bookmarks themselves
     _parser.add_argument('file',
                          type=str,
+                         nargs="?",
                          metavar="File/Alias",
                          help="The file/alias that will be used")
 
@@ -56,73 +56,71 @@ def argParser() -> argparse.ArgumentParser:
                                     action="store_true",
                                     help="Lists every alias in current environment")
 
-    # TODO: Add "run" (run file)/"edit" (open code editor or whatever!)
-    # (if that can actually be implemented)
-
     return _parser
 
 
 def echoFile(FilePath: Path) -> None:
     import mmap
     from shutil import copyfileobj
-    # TODO: Maybe implementing memory mapping to other places in this script would be good
     with open(FilePath, 'rb') as f:
         with mmap.mmap(f.fileno(), 0, access=mmap.ACCESS_READ) as s:
             copyfileobj(s, sys.stdout.buffer)
 
 
 def main() -> None:
-    # Variable declaration
+    # Create json data file if does'nt exist (or is empty) and reference it in main_data_file
     currosslash = "\\" if (os.name == "nt") else "/"
     main_data_file = Path(
         f"{Path(__file__).parent.absolute()}{currosslash}data.json")
-    # TODO: Implement "config" file with toml
-    # main_cfg_file = smth
     del currosslash
-    if not main_data_file.exists():
+    if not main_data_file.exists() or os.stat(main_data_file).st_size == 0:
         with open(main_data_file, 'w') as f:
             f.write("{}")
 
-    # Start handling args
-    args = argParser().parse_args()
+    # Parse the args!
+    parser = argParser()
+    args = parser.parse_args()
     env = args.env
 
+    # Check if env is actually in there, if not, create new one.
     if isinstance(args.env, str):
-        # Check if env is actually in there, if not, create new one.
         with open(main_data_file, "r+") as f:
             full_data = json.load(f)
-            if args.env not in full_data:
-                full_data[args.env] = {}
+            if env not in full_data:
+                full_data[env] = {}
                 f.seek(0)
                 json.dump(full_data, f, indent=4)
                 f.truncate()
 
-    # If args.main is an alias, set it to it's path
-    with open(main_data_file, "r") as f:
-        full_data = json.load(f)
+    # If args.file is an alias, set it to the alias'es filepath
+    # And make a variable called filepath
     original_alias = args.file
-    if args.file in full_data[env]:
-        args.file = full_data[env][args.file]
-    del full_data
-    filepath = Path(args.file).absolute()
-
-    # Handle arguments properly now
-    if args.list == True:
-        # Literally just print everything under XYZ env
-        with open(main_data_file, "r+") as f:
+    if not args.file == None:
+        # Share this single "read" call for other every argument
+        with open(main_data_file, "r") as f:
             full_data = json.load(f)
+        if args.file in full_data[env]:
+            args.file = Path(full_data[env][args.file])
+        else:
+            args.file = Path(args.file)
+        # If it somehow isn't an absolute path, make it absolute anyways
+        filepath = args.file.absolute()
+
+    # Mutulally exclusive arguments:
+
+    # Print alias'es folder, if not, print all of env's variables
+    if args.list == True:
+        if args.file != None:
+            if filepath.is_dir():
+                print(*os.listdir(filepath))
+                sys.exit()
         print(f"These are the aliases within env '{env}':")
         for key, val in full_data[env].items():
             print(f'"{key}": "{val}"')
         sys.exit(0)
 
-    if args.path == True:
-        if not filepath.exists():
-            print("Not a valid file or directory")
-            sys.exit(1)
-        print(filepath)
-
-    if args.output == True:
+    # Depending on which kind of file this is, output something to stdout
+    if args.output == True and not args.file == None:
         if not filepath.exists():
             print("Not a valid file or directory")
             sys.exit(1)
@@ -132,24 +130,22 @@ def main() -> None:
             print(*os.listdir(filepath))
         sys.exit(0)
 
-    if not args.add == None:
-        alias = args.add
+    # Adds keyword + file as an json key if it doesnt exist,
+    # otherwise, update existing one
+    if not args.add == None and not args.file == None:
         if not filepath.exists():
             print("Not a valid filepath, exiting.")
             sys.exit(1)
-        with open(main_data_file, "r") as f:
-            full_data = json.load(f)
-        # If alias already exists in env
-        update_alias = False
-        if alias in full_data[env]:
-            update_alias = True
-        full_data[env][alias] = f"{filepath}"
 
+        alias = args.add
+        full_data[env][alias] = str(filepath)
         with open(main_data_file, "w") as f:
             f.seek(0)
             json.dump(full_data, f, indent=4)
             f.truncate()
 
+        # If alias already exists in env
+        update_alias = True if alias in full_data[env] else False
         if update_alias:
             print(f"Alias '{alias}' in env '{env}' updated successfully!")
         else:
@@ -157,24 +153,29 @@ def main() -> None:
                 f"Alias '{alias}' added to environment '{env}' successfully!")
         sys.exit(0)
 
-    if args.remove == True:
+    # Removes alias from data file
+    if args.remove == True and not args.file == None:
         with open(main_data_file, "r+") as f:
-            full_data = json.load(f)
             if not original_alias in full_data[env]:
                 print(f'Alias {original_alias} not found under env "{env}"')
                 sys.exit(1)
-            del full_data[env][original_alias]
-            f.seek(0)
-            json.dump(full_data, f, indent=4)
-            f.truncate()
-        print(f'Alias {original_alias} sucessfully removed from env "{env}"!')
+            else:
+                del full_data[env][original_alias]
+                f.seek(0)
+                json.dump(full_data, f, indent=4)
+                f.truncate()
+                print(
+                    f'Alias {original_alias} sucessfully removed from env "{env}"!')
+                sys.exit(0)
+
+    # if there's no file and there was no other option, print help
+    # also, if nothing else than the file is specified, run --path
+    if args.file == None:
+        parser.print_help()
+        sys.exit(0)
+    else:
+        print(filepath)
 
 
 if __name__ == '__main__':
-    # Weird workaround, but works pretty well!!
-    if len(sys.argv) <= 1:
-        sys.argv += "."
-    if "-" in sys.argv[1]:
-        sys.argv += "."
-        sys.argv = [*sys.argv, sys.argv.pop(1)]
     main()
